Renaming directory to conform to standards
