+++ pages/3_Compare_Cases.py
@@ -2,9 +2,9 @@
 
 import streamlit as st
 
-from sibylapp.compute.context import get_term
-from sibylapp.view import entity_difference
-from sibylapp.view.utils import display, filtering, setup
+from sibylapp2.compute.context import get_term
+from sibylapp2.view import entity_difference
+from sibylapp2.view.utils import display, filtering, setup
 
 setup.setup_page(return_row_ids=True)
