+++ sibylapp2/compute/contributions.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import streamlit as st
 
-from sibylapp.compute import api, entities
+from sibylapp2.compute import api, entities
 
 
 @st.cache_data(show_spinner="Computing contribution scores...")
