+++ sibylapp2/compute/example_based.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from sibylapp.compute import api
+from sibylapp2.compute import api
 
 
 @st.cache_data
