+++ sibylapp2/compute/model.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from sibylapp.compute import api, entities
+from sibylapp2.compute import api, entities
 
 
 @st.cache_data(show_spinner="Getting model predictions...")
