+++ sibylapp2/view/feature_importance.py
@@ -1,9 +1,9 @@
 import streamlit as st
 
-import sibylapp.view.utils.filtering
-from sibylapp.compute import importance
-from sibylapp.compute.context import get_term
-from sibylapp.view.utils import helpers
+import sibylapp2.view.utils.filtering
+from sibylapp2.compute import importance
+from sibylapp2.compute.context import get_term
+from sibylapp2.view.utils import helpers
 
 
 @st.cache_data
@@ -24,7 +24,7 @@ def format_importance_to_view(importance_df):
 def view():
     to_show = format_importance_to_view(importance.compute_importance())
     to_show = to_show.sort_values(by="Importance Value", axis="index", ascending=False)
-    to_show = sibylapp.view.utils.filtering.process_options(to_show)
+    to_show = sibylapp2.view.utils.filtering.process_options(to_show)
     helpers.show_table(to_show.drop("Importance Value", axis="columns"), key="importance")
     return to_show["Feature"]
