+++ sibylapp2/view/global_contributions.py
@@ -3,10 +3,10 @@ import pandas as pd
 import streamlit as st
 from pyreal.visualize import strip_plot
 
-from sibylapp.compute import contributions
-from sibylapp.compute.context import get_term
-from sibylapp.view.utils import filtering, helpers
-from sibylapp.view.utils.helpers import show_legend
+from sibylapp2.compute import contributions
+from sibylapp2.compute.context import get_term
+from sibylapp2.view.utils import filtering, helpers
+from sibylapp2.view.utils.helpers import show_legend
 
 
 @st.cache_data(show_spinner="Generating summary plot...")
