+++ sibylapp2/view/similar_entities.py
@@ -1,10 +1,10 @@
 import pandas as pd
 import streamlit as st
 
-from sibylapp.compute import example_based
-from sibylapp.compute.context import get_term
-from sibylapp.view.utils import helpers
-from sibylapp.view.utils.filtering import process_options
+from sibylapp2.compute import example_based
+from sibylapp2.compute.context import get_term
+from sibylapp2.view.utils import helpers
+from sibylapp2.view.utils.filtering import process_options
 
 
 def format_similar_entities(x, y):
