+++ sibylapp2/view/utils/formatting.py
@@ -1,5 +1,5 @@
-from sibylapp.compute.context import get_term
-from sibylapp.view.utils import helpers
+from sibylapp2.compute.context import get_term
+from sibylapp2.view.utils import helpers
 
 
 def format_single_contributions_df(df, show_number=False):
